+++ src/interface/src/app/app-routing.module.ts
@@ -15,11 +15,8 @@ import { MapComponent } from './map/map.component';
 import { CreateScenariosComponent } from './plan/create-scenarios/create-scenarios.component';
 import { PlanComponent } from './plan/plan.component';
 import { PasswordResetComponent } from './password-reset/password-reset.component';
-import {
-  AuthGuard,
-  ValidationResolver,
-  passwordResetTokenResolver,
-} from './services';
+import { AuthGuard, ValidationResolver } from './services';
+import { passwordResetTokenResolver } from './services/password-reset';
 import { SignupComponent } from './signup/signup.component';
 import { RedirectGuard } from './services/redirect.guard';
