+++ src/interface/src/app/password-reset/confirmation-dialog/confirmation-dialog.component.html
@@ -1,5 +1,5 @@
 <mat-dialog-content>
-  <img src="{{ checkUrl }}" alt="" />
+  <img src="{{ checkImageUrl }}" alt="" />
   <h3
     class="gmat-headline-3"
     i18n="
