+++ src/interface/src/app/password-reset/password-reset.component.html
@@ -25,7 +25,7 @@
         <button
           mat-stroked-button
           color="secondary"
-          type="cancel"
+          type="button"
           class="reset-form-button"
           (click)="cancel()">
           Cancel
